### 2021-09-10 #3712 @whedon
[PRE REVIEW]: DRSP-Sim: A Simulator for Ride-Sharing with Pooling: Joint Matching, Pricing, Route Planning, and Dispatching
**Submitting author:** @marina-haliem (<a href="http://orcid.org/0000-0002-9782-6591">Marina Haliem</a>)
**Repository:** <a href="https://github.com/marina-haliem/Dynamic-RideSharing-Pooling-Simulator" target ="_blank">https://github.com/marina-haliem/Dynamic-RideSharing-Pooling-Simulator</a>
**Version:** v1.1
