### 2021-09-23 #3761 @whedon
[REVIEW]: DRSP-Sim: A Simulator for Ride-Sharing with Pooling: Joint Matching, Pricing, Route Planning, and Dispatching
**Submitting author:** @marina-haliem (<a href="http://orcid.org/0000-0002-9782-6591">Marina Haliem</a>)
**Repository:** https://github.com/marina-haliem/Dynamic-RideSharing-Pooling-Simulator
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
