### 2021-09-23 #3761 @martinfleis
I am copying here https://github.com/openjournals/joss-reviews/issues/3712#issuecomment-924912823 from @RafalKucharskiPK for completeness. 👇

Dear @marina-haliem
 I cloned the repo, downloaded the data (from https://purr.purdue.edu/publications/3843/1), created conda env, installed most of dependencies and run `simulator_driver.py`.

 Minor issues that I managed to overcome:
1. you require pandas==0.22 - which is ancient - 2017(!), and it clashed with many of my other packages, I used lates pandas and so far it went OK.
2. The `db.sqlite3` path is your local path, I changed it to "data\db":
![image](https://user-images.githubusercontent.com/20555451/134348441-ffc89cf0-d2f4-4f18-8d3c-735cb3fc9ba7.png)

3. The default simulation takes 8 days (in simulation) one day takes ca. one hour on my laptop - way too long for a quickstart, please provide a toy example for first users.

Nonetheless it did not work for me, after a single day the error occured. Then I changed to one day in config, and again after a day passed I receive:
```
Traceback (most recent call last):
  File "/Users/rkucharski/Documents/GitHub/Dynamic-RideSharing-Pooling-Simulator/simulator_driver.py", line 159, in <module>
    Sim_experiment.simulator.step()
  File "/Users/rkucharski/Documents/GitHub/Dynamic-RideSharing-Pooling-Simulator/simulator/simulator.py", line 71, in step
    if vehicle.agent_type == agent_codes.dqn_agent:
AttributeError: 'Vehicle' object has no attribute 'agent_type'

Process finished with exit code 1
```
 so please create a stable version to reproduce.

PS. On the main github page I am missing quick walkthrough tutorial for first users, e.g. now I first see how to set 'OSRM' which happens to be not needed, but only later I see where is the data bundle to run simulations.

Good luck!
