+++ requirements.txt
@@ -2,7 +2,9 @@ psycopg2-binary==2.7.4
 SQLAlchemy>=1.2.2
 PyYAML>=3.12
 tensorflow==1.15.0
-pandas==0.22.0
-polyline==1.3.2
+pandas==1.2.0
+keras==2.3.1
+polyline>=1.3.2
 requests>=2.19.1
+numpy>=1.18.1
 scikit-image==0.14.0
