### 2021-09-30 299cd55
Updated Getting Started section
