### 2021-09-30 ed2517c
Updated Getting Started section
