### 2021-09-30 #3761 @marina-haliem
Hi @RafalKucharskiPK,

Thank you for your comments. I have updated the requirements list, and the default db_dir path. Also, I've updated the Getting Started section, highlighting how to run a quick start of 1 day. Apologies regarding this error, there was a minor bug that I fixed. I have conducted some additional testing, and I believe it should run smoothly now for you.

Please, let us know if you come across any further issues.

Thank you,
Marina
