+++ parse_results.py
@@ -7,17 +7,17 @@
 
 # Import LogAnalyzer objects
 from tools.log_analyzer import *
+from config.settings import DEFAULT_LOG_DIR
 
 # and other relevant stuff...
 import matplotlib.pyplot as plt
 
 # #### !IMPORTANT: Specify directory and log filenames here 
 # Note that the provided names (below) are default names. They do not have to be changes unless you decided to rename files from multiple experiments.
-log_dir_path = "./baselines/DARM/sim/"
-vehicle_log_file = "vehicle.log"
-customer_log_file = "customer.log"
-score_log_file = "score.log"
-summary_log_file = "summary.log"
+vehicle_log_file = "/sim/vehicle.log"
+customer_log_file = "/sim/customer.log"
+score_log_file = "/sim/score.log"
+summary_log_file = "/sim/summary.log"
 
 
 # Invoke LogAnalyzer Object
@@ -26,10 +26,10 @@ l = LogAnalyzer()
 
 # #### Exploring dataframes of each of the logs
 # Loading all the different logs as pandas dataframes
-summary_df = l.load_summary_log(log_dir_path)
-vehicle_df = l.load_vehicle_log(log_dir_path)
-customer_df = l.load_customer_log(log_dir_path)
-score_df = l.load_score_log(log_dir_path)
+summary_df = l.load_summary_log(DEFAULT_LOG_DIR)
+vehicle_df = l.load_vehicle_log(DEFAULT_LOG_DIR)
+customer_df = l.load_customer_log(DEFAULT_LOG_DIR)
+score_df = l.load_score_log(DEFAULT_LOG_DIR)
 
 print(summary_df.describe())
 
@@ -48,26 +48,18 @@ df = l.get_customer_waiting_time(customer_df)
 print(df.head())
 
 
-# #### Generating plots of summary logs
-DARM = "./baselines/DARM/sim/"
-DPRS = "./baselines/DPRS/sim/"
-Deep_Pool = "./baselines/Deep_Pool/sim/"
-ODA = "./baselines/ODA/sim/"
-Dummy_RS = "./baselines/Dummy_RS/sim/"
-Dummy_None = "./baselines/Dummy_None/sim/"
+##### Generating plots of summary logs
+## You can have more than one experiment that you want to compare, you can pass their relevant paths here to the first
+## list argument
+summary_plots = l.plot_summary([DEFAULT_LOG_DIR], ["Number of Accepted Requests", "Average Travel Distance",
+												   "Occupancy Rate of Vehicles"], plt)
+summary_plots.savefig(DEFAULT_LOG_DIR+"/Summary.pdf", bbox_inches = 'tight')
+summary_plots.show()
 
-# summary_plots = l.plot_summary([DARM, DPRS, Deep_Pool, ODA, Dummy_RS, Dummy_None], ["Number of Accepted Requests",
-# 																"Number of Rejected Requests", "Occupancy "
-# 																										"Rate of "
-# 																							   "Vehicles"], plt)
-# summary_plots.savefig("./baselines/DARM/Summary.pdf", bbox_inches = 'tight')
-# summary_plots.show()
-
-# #### Generating plots of relevant experiment metrics
-plt, df = l.plot_metrics([log_dir_path], ["Profit", "Cruising Time", "Occupancy Rate","Waiting Time", "Travel "
-																								"Distance"],plt)
-plt.savefig("./baselines/DARM/Metrics.pdf", bbox_inches = 'tight')
-plt.show()
+##### Generating plots of relevant experiment metrics
[…]
