### 2021-10-23 #3761 @marina-haliem
Thanks, @RafalKucharskiPK, for your comments.

- I am not able to replicate this issue. I've run this project so many times but never encountered a similar issue. The only thing is that the logs will need to be cleared in case of re-running the experiments, or renamed differently to distinguish different experiments using 'DEFAULT_LOG_DIR' in config/settings.py. If there is a specific error that you encountered, please let us know and we will try to resolve it.
- I've updated parse_results.py and log_analyzer.py to avoid the confusion of the parameters. By default, the parse_results.py will provide a tool to plot the results of the default log directory specified in 'DEFAULT_LOG_DIR'. Since the user can decide to compare any number of experiments without any constraints, it is left up to them to specify their relevant paths to generate relevant plots. I have also updated the relevant description in the README file for more clarity.

Thank you,
Marina
