### 2021-12-18 #3761 @martinfleis
Hi @marina-haliem ,

First, apologies for the long review process. It is not always possible to make it swift, especially with more common unexpected issues these days.

I have started my review of DRSP-Sim and ticked relevant boxes in the checklist above. However, I cannot at the moment tick the rest, in some cases due to missing components of the submission, in others due to (I assume) technical limitations on my side.

**Installation**

- I am running an M1 mac laptop on which I am unable to install dependencies pinned in the `requirements.txt` file. My assumption is that it is caused by the chip architecture so I’ll try to do the same on an intel-based machine later.
- Due to this I wasn’t able to check Functionality and Performance yet.

**Example usage** and **Functionality documentation**

- I believe that the documentation (or a lack of it) is the major issue here. The ReadMe contains very rough set of instructions on how to run the script but there’s no other documentation.
- Functions of the package are not listed anywhere. You [need](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) API documentation.
- The functions are not documented even within the source code, so looking inside it doesn’t help. There are no docstrings.

**Automated tests**

- There are no [automated tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) either. There is no way to objectively assess whether the software works as intended.

**Community guidelines**

- I have ticked community guidelines but I believe that it would benefit if you could expand it. There’s no developer documentation at the moment, so it is unclear how to create a dev environment and what needs to be done to prepare a PR (do I need to follow some code style, write tests, write documentation?).

Apart from these points from the checklist, I also have a couple of additional questions and a few minor comments.

- What was the role of your co-authors? There are no commits from them, can you explain their contribution?
- The links to the git repository in the Readme point to the old Purdue address. Can you update them to point to the current Github repository?
- Formatting of the code snippets in the Readme is broken in some cases (code is formatted as a standard text).
- How can I use the library to do a simulation of, say, Prague? What is the required data input? How to create it? All that should be documented if you want it to be reusable.
- If you want DRSP-Sim to be usable in the long term, you should avoid pinned versions of your dependencies. As @RafalKucharskiPK pointed above, your list breaks existing environments and doesn’t install on Arm-based machines.

Some of these points (documentation, tests) will undoubtedly require some work but you will need to address them to meet the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#).

Let me know if there’s anything unclear in my comments.
