### 2022-01-26 #3761 @RafalKucharskiPK
Dear Editor, Dear Authors @marina-haliem ,

sorry for the delayed process and sorry that you did not manage to collect more reviewers.
I believe we may divide the review process into two separate parts:
1. Paper
2. Software

Since the software is well reviewed by @martinfleis and I had similar issues at the installation, I will focus more on the **paper**:

1. I am missing the use-cases. You well review the existing alternative models and highlight novelty, but I am still missing use-cases
2. The functionality of modules used in the architecture is missing. Why do we have them - how can we use them (parameterise) - and what are the use-cases for applying them in studies.
3. All in all the paper is very brief and does not give an overview of what is the software - it focuses a bit why is it better than other similar ones, but misses the point on what it really is.
4. The structure is misleading for me: you have Benchmarks as the section and then under Benchmarks you have details of pricing and mathing algorithms in subsections - are they benchmarks?

Please try to improve the clarity and completeness of the paper. By the way the issues are very similar to the issues found in the software description.

**Software**:
To summarize my previous issues - I am missing a walkthrough tutorial: starting from uses-case scenario (what do we want to simulate), through set-up, input data - through running and up to interpreation of the results. Without this I'd miss the whole story.

Hope this helps improve the paper!
Rafal Kucharski
