### 2022-02-15 #3761 @martinfleis
@RafalKucharskiPK In the very top post in this issue is an automatically crated review checklist prepared for you. We use it to ensure that the paper fulfils all the JOSS requirements and wait until all boxes are ticked before accepting it. You should be able to edit that post and tick boxes you think the current version of the software and the paper meets.
If you are unable to do so, I might need to re-invite you to join the JOSS org.
