+++ simulator/models/vehicle/vehicle_behavior.py
@@ -8,6 +8,7 @@ class VehicleBehavior(object):
     def step(self, vehicle, timestep):
         pass
 
+    # Vehicle drives according to set route and speed
     def drive(self, vehicle, timestep):
         route = vehicle.get_route()  # Sequence of (lon, lat)
         speed = vehicle.get_speed()
@@ -15,8 +16,6 @@ class VehicleBehavior(object):
         rlats, rlons = zip(*([vehicle.get_location()] + route))  # New vehicle location after driving this route
         step_dist = geoutils.great_circle_distance(rlats[:-1], rlons[:-1], rlats[1:],
                                                    rlons[1:])  # Get distcnace in meters
-        # print(step_dist)
-        # print(type(step_dist))
         vehicle.state.travel_dist += dist_left
 
         for i, d in enumerate(step_dist):
@@ -39,7 +38,7 @@ class Idle(VehicleBehavior):
 
 
 class Cruising(VehicleBehavior):
-    # Updated remaining time to destination, if arrived states changes to parking
+    # Updated remaining time to destination, if arrived: state changes to parking
     def step(self, vehicle, timestep):
         arrived = vehicle.update_time_to_destination(timestep)
         if arrived:
@@ -48,7 +47,7 @@ class Cruising(VehicleBehavior):
 
         self.drive(vehicle, timestep)
 
-# NEEDS TO BE UPDATED (Dropoff one customer at a time) + Update location and new route
+# Dropoff one customer at a time + Update location and new route
 class Occupied(VehicleBehavior):
     available = False
     # Updated remaining time to destination, if arrived customer gets off
@@ -67,13 +66,7 @@ class Occupied(VehicleBehavior):
             vehicle.state.accept_new_request = False
 
         if arrived:
-            # customer = vehicle.dropoff()
-            # customer.get_off()
-            # print(len(vehicle.current_plan), vehicle.current_plan)
-            # print(vehicle.ordered_pickups_dropoffs_ids)
-            id = vehicle.ordered_pickups_dropoffs_ids.pop(0)  # CHNAGE THIS LIST TO QUEUE, POP HERE AND HEAD TO
-            # NEXT
-            # print("vid: ", vehicle.get_id(), " -> pop: ", id)
+            id = vehicle.ordered_pickups_dropoffs_ids.pop(0)  # POP HERE AND HEAD TO NEXT
             customer = simulator.models.customer.customer_repository.CustomerRepository.get(id)
             if vehicle.pickup_flags.pop(0) == 1:
                 # print(vehicle.get_id(), "Occupied -> Pickup")
@@ -85,7 +78,7 @@ class Occupied(VehicleBehavior):
         self.drive(vehicle, timestep)
 
 
-# NEEDS TO BE UPDATED (Pickup one customer at a time) + Update location and new route
+# Pickup one customer at a time + Update location and new route
 class Assigned(VehicleBehavior):
     available = False
     # Updated remaining time to destination, if arrived, update customer ID and picks him up
@@ -99,21 +92,8 @@ class Assigned(VehicleBehavior):
             # print("Vid: ", vehicle.get_id(), "Arrived ", arrived)
 
         if arrived:
-            # if FLAGS.enable_pooling:
-                # print("Assigned, pooling!")
-            # print(vehicle.ordered_pickups_dropoffs_ids)
-            id = vehicle.ordered_pickups_dropoffs_ids.pop(0)    # CHNAGE THIS LIST TO QUEUE, POP HERE AND HEAD TO
-            # NEXT
-            # print("vid: ", vehicle.get_id(), "Loc: ", vehicle.get_location(), " -> pop: ", id)
-            # print("Arrived: Vehicle Info", vehicle.to_string())
-            # print("Customer ids:", ids)
[…]
