### 2022-04-01 5a4ec75
Update vehicle_repository.py
