+++ simulator/models/vehicle/vehicle_repository.py
@@ -11,11 +11,13 @@ class VehicleRepository(object):
 
     @classmethod
     def populate(cls, vehicle_id, location, type):
+        # populate vehicles over the city maps.
         state = VehicleState(vehicle_id, location, type)
         cls.vehicles[vehicle_id] = Vehicle(state)
 
     @classmethod
     def get_all(cls):
+        # returns a list of all vehicles in repository.
         return list(cls.vehicles.values())
 
     @classmethod
@@ -24,6 +26,7 @@ class VehicleRepository(object):
 
     @classmethod
     def get_states(cls):
+        # get all vehicles' state attributes from repository.
         states = [vehicle.get_state() for vehicle in cls.get_all()]
         cols = VehicleState.__slots__[:]
         df = pd.DataFrame.from_records(states, columns=cols).set_index("id")    # Creating DF with all attributes in Vehicle State
@@ -38,4 +41,5 @@ class VehicleRepository(object):
 
     @classmethod
     def delete(cls, vehicle_id):
+        # delete when vehicle leaves the market.
         cls.vehicles.pop(vehicle_id)
