### 2022-04-02 c6cc0e4
Added Functionalities
Added Functionalities to the modular architecture diagram
