### 2022-04-02 #3761 @marina-haliem
Thank you for your comments and feedback.

**Reviewer 1** @martinfleis :

**Installation**
Would you elaborate on what issues are facing with this? I have personally run this project on a Mac laptop without any problems.

**Example usage and Functionality documentation**

1. I have documented the functionalities within the source code for more clarity as well as adding a thorough explanation of each module's functionalities in the paper under the (Simulator Architecture and Functionalities Section).
2. The modular code architecture diagram has also been updated to include all of the core functionalities of each sub-module.

**Testing**
For testing, we have provided manual steps in the Readme document that can be followed along with a sample expected output to check the expected functionality of the software.

**Community guidelines**
The developer guidelines are provided in the Readme including how to set up a dev environment, run the simulator, and generate plots to interpret the logs. No special code style is needed, just regular python code.

1. This codebase was initially developed over the old Purdue Github link, co-authors commits would have shown there. This Github link was mainly created for this submission.
2. Repository link updated in the Readme. We intend to move the repository to another Github link in the future.
Formatting of the code snippets in the Readme is broken in some cases (code is formatted as a standard text).
3. This can be found under the (Data Generation) section in the Readme. If you would like to simulate any other city map (other than NY), you basically follow the same preprocessing steps listed there, but using the osm data of the required city as well as its corresponding trip data as well to generate the discretized city map.
4. I've updated parse_results.py and log_analyzer.py to avoid confusion of the parameters as @RafalKucharskiPK pointed out above. By default, the parse_results.py will provide a tool to plot the results of the default log directory specified in 'DEFAULT_LOG_DIR'. Since the user can decide to compare any number of experiments without any constraints, it is left up to them to specify their relevant paths to generate relevant plots. I have also updated the relevant description in the README file for more clarity.  I am not sure if Arm-based machines processors would cause any changes, but my team and I have successfully ran this project several times.

Thank you,
Marina
