### 2022-04-02 #3761 @marina-haliem
Thank you for your comments and feedback.

Reviewer 2 @RafalKucharskiPK :

- For use cases, I relied on the Readme for that since the paper is limited in the number of pages. In the Readme, we have provided a thorough step-by-step explanation for simulator setup, input data, running the simulator, storing logs, and interpreting them by generating plots. Besides, anyone in the area can utilize the simulator for their own purposes, customizing the matching, pricing, and dispatching policies according to their research needs.
- For the functionality of modules used in the architecture, I have added a thorough explanation of each module's core functions in the paper under the (Simulator Architecture and Functionalities Section). The modular code architecture diagram has also been updated to include all of the core functionalities of each sub-module. For each module, our simulator provides one or two benchmarks for usage flexibility, but the simulator is meant to be utilized by any researcher in the area to plug in their own algorithms and compare the results.
- We emphasize the purpose of the simulator in the paper under the (Statement of need) section, stating that "Also, algorithms of customer-vehicle matching, route planning, pricing, and dispatching have been developed; however, these sub-problems tend to be studied separately and a complete integrated simulator that considers pooling is still lacking. We present the first RS simulator that integrates all of these together, provides a baseline for each and allows developers to plug-and-play algorithms to tackle any of the sub-problems while defaulting the rest to the provided benchmark. This, in turn, will greatly facilitate cross-study evaluations of any existing ridesharing algorithms, while providing guidance for future research."
- Under the (Benchmarks) section, we provide one or two benchmark algorithms for each sub-module (matching, pricing, and dispatching). This is what the simulator defaults to unless other algorithms are provided by the user for one or all of these sub-modules.

**Software:**
In the Readme, under the (Getting Stated) section, we provide a step-by-step explanation of creating a dev environment, setting up the simulator parameters that provide wide flexibility, loading and defining the input data, running the simulator, storing logs, post-processing and generating plots to interpret these logs. We also provide sample output of what plots would look like, if it were to compare two algorithms and if it were to visualize a summary of metrics for one algorithm.

Thank you,
Marina
