### 2022-04-09 #3761 @martinfleis
Hi @marina-haliem,

thank you for all the changes and a thorough response!

> Would you elaborate on what issues are facing with this? I have personally run this project on a Mac laptop without any problems.

Yes. When following the installation instructions, after cloning the repository, running

```
pip install -r requirements.txt
```
in a fresh environment fails with the following error message.

```
Collecting psycopg2-binary==2.7.4
  Using cached psycopg2-binary-2.7.4.tar.gz (426 kB)
  Preparing metadata (setup.py) ... error
   command: /Users/martin/mambaforge/bin/python3.9 -c 'import io, os, sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-install-yvq80uz5/psycopg2-binary_d9f6f120662c49d7899e55e339bbb9ea/setup.py'"'"'; __file__='"'"'/private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-install-yvq80uz5/psycopg2-binary_d9f6f120662c49d7899e55e339bbb9ea/setup.py'"'"';f = getattr(tokenize, '"'"'open'"'"', open)(__file__) if os.path.exists(__file__) else io.StringIO('"'"'from setuptools import setup; setup()'"'"');code = f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-pip-egg-info-ku_fu6e_
       cwd: /private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-install-yvq80uz5/psycopg2-binary_d9f6f120662c49d7899e55e339bbb9ea/
  Complete output (23 lines):
  running egg_info
  creating /private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-pip-egg-info-ku_fu6e_/psycopg2_binary.egg-info
  writing /private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-pip-egg-info-ku_fu6e_/psycopg2_binary.egg-info/PKG-INFO
  writing dependency_links to /private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-pip-egg-info-ku_fu6e_/psycopg2_binary.egg-info/dependency_links.txt
  writing top-level names to /private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-pip-egg-info-ku_fu6e_/psycopg2_binary.egg-info/top_level.txt
  writing manifest file '/private/var/folders/2f/fhks6w_d0k556plcv3rfmshw0000gn/T/pip-pip-egg-info-ku_fu6e_/psycopg2_binary.egg-info/SOURCES.txt'

  Error: pg_config executable not found.

  pg_config is required to build psycopg2 from source.  Please add the directory
  containing pg_config to the $PATH or specify the full executable path with the
  option:

      python setup.py build_ext --pg-config /path/to/pg_config build ...

  or with the pg_config option in 'setup.cfg'.

  If you prefer to avoid building psycopg2 from source, please install the PyPI
  'psycopg2-binary' package instead.

  For further information please check the 'doc/src/install.rst' file (also at
  <http://initd.org/psycopg/docs/install.html>).

WARNING: Discarding https://files.pythonhosted.org/packages/77/09/4991fcd9a8f4bea1ee3948e1729fa17c184d25bd10809bacc143626361b9/psycopg2-binary-2.7.4.tar.gz#sha256=de4f88f823037a71ea5ef3c1041d96b8a68d73343133edda684fd42f575bd9d7 (from https://pypi.org/simple/psycopg2-binary/). Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.
ERROR: Could not find a version that satisfies the requirement psycopg2-binary==2.7.4 (from versions: 2.7.4, 2.7.5, 2.7.6, 2.7.6.1, 2.7.7, 2.8, 2.8.1, 2.8.2, 2.8.3, 2.8.4, 2.8.5, 2.8.6, 2.9, 2.9.1, 2.9.2, 2.9.3)
ERROR: No matching distribution found for psycopg2-binary==2.7.4
```

The version 2.7.4 you, for some reason, require is quite old (Feb 2018) and I believe that is causing the issue when you try to install it on Macs with M1 chips. There were no dependencies built for these chips back then.
