+++ requirements.txt
@@ -1,10 +1,9 @@
-psycopg2-binary==2.7.4
 SQLAlchemy>=1.2.2
-PyYAML>=3.12
 tensorflow==1.15.0
-pandas==1.2.0
+pandas>=1.2.0
 keras==2.3.1
-polyline>=1.3.2
+polyline>=1.4.0
 requests>=2.19.1
-numpy>=1.18.1
-scikit-image==0.14.0
+numpy>=1.18.5
+pyyaml==5.3.1
+scikit-image==0.14.2
