+++ requirements.txt
@@ -5,5 +5,5 @@ keras==2.3.1
 polyline>=1.4.0
 requests>=2.19.1
 numpy>=1.18.5
-pyyaml==5.3.1
+pyyaml==5.4
 scikit-image==0.14.2
