### 2022-06-16 #3761 @martinfleis
@marina-haliem It is waiting for the second round of revisions, sorry for delays.

@RafalKucharskiPK can you have a look at the response above and update the checklist in [the first post](https://github.com/openjournals/joss-reviews/issues/3761#issue-1005904121)?
