### 2022-07-19 marina-haliem/Dynamic-RideSharing-Pooling-Simulator#3 @martinfleis
Installation on M1 Macs
Hi @marina-haliem,

I am still trying to install the requirements as per the instructions on my M1 MacBook but I don't think it will be possible under current circumstances.

```
(drsp-sim) martin@Martins-MacBook-Air Dynamic-RideSharing-Pooling-Simulator % pip install -r requirements.txt
Collecting SQLAlchemy>=1.2.2
  Downloading SQLAlchemy-1.4.39.tar.gz (8.2 MB)
     |████████████████████████████████| 8.2 MB 2.1 MB/s
  Preparing metadata (setup.py) ... done
ERROR: Could not find a version that satisfies the requirement tensorflow==1.15.0 (from versions: none)
ERROR: No matching distribution found for tensorflow==1.15.0
```

The requirements pin `tensorflow==1.15.0` which was released on Oct 16, 2019 when no M1 chips existed, so there is no wheel for this architecture.

Unless this pin is lifted, there is no way you will be able to run the code on M1 chips. conda is also unable to resolve such old versions.

```
Encountered problems while solving:
  - nothing provides requested tensorflow 1.15.0
  - nothing provides requested keras 2.3.1
  - nothing provides requested pyyaml 5.4
  - nothing provides requested scikit-image 0.14.2
```

xref https://github.com/openjournals/joss-reviews/issues/3761
