### 2022-10-15 #3761 @martinfleis
Use `@editorialbot generate my checklist` command. I didn't realise this review is pre-editorialbot.
