### 2022-10-15 #3761 @jGaboardi
@marina-haliem @martinfleis

Before continuing with my review, I would like to raise two fairly large points (IMHO), that need to be addressed:

* **software**
  * One of the overarching goals of JOSS, and, similar outlets, is to advance the concept of (open source) software as a scientific artifact. While `DRSP-Sim` is technically open source, it is not accessible for generalized use (e.g. it can't be installed and is not available via `pip` or `conda-forge`). Moreover, as @martinfleis pointed out, there is no documentation site and there are no docstrings to speak of. Where [there are docstrings](https://github.com/marina-haliem/Dynamic-RideSharing-Pooling-Simulator/blob/master/common/geoutils.py#L30-L46), they are not formatted in a standard way.
* **paper**
  * The paper is well written, but 7 of the 17 references are to works involving authors of this manuscript. With a quick search there seems to be other related open software that should be cited. For example, [`MaaSSim`](https://github.com/RafalKucharskiPK/MaaSSim) by a previous reviewer.
