### 2022-11-15 #3761 @martinfleis
@marina-haliem Can you respond to comments raised by @jGaboardi above? Please keep in mind that the review process of JOSS is different than in a standard journal and assumes direct communication rather than fixed rounds of review-response cycle. Thank you!
